Beginning database interactions
This update to the project rectifies the 30 instances of the number 2, It also begins the implementation of SQLite usage for database entries.  The database_helper.dart is the implementation of that database structure and connection. Not all fields are involved at the moment, but there is no retrieval within the app at this point.
With the client/wife's input, the slide bar now has demarcations going from 0 to 10.